Revert "feature: process calling subset_shape."
This reverts commit 4dc92ddc5ceacbcec7d3f1cfc911766b80367638.
